Revert "Extract PendingOpSlots and split slow-path token generation between event-loop and off-loop submissions"
This reverts commit 71f78223bb734ec100470cb5e35bf371d81f86a7.
