+++ transport-classes-io_uring/src/main/java/io/netty/channel/uring/IoUringIoHandler.java
@@ -42,6 +42,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicLong;
 
 import static java.lang.Math.max;
 import static java.lang.Math.min;
@@ -567,14 +568,14 @@ public final class IoUringIoHandler implements IoHandler {
                 }
                 return packedSeq;
             }
-            boolean inEventLoop = executor.isExecutorThread(Thread.currentThread());
-            long seq = pendingOps.nextToken(inEventLoop);
-            if (inEventLoop) {
+            long seq = pendingOps.nextToken();
+            if (executor.isExecutorThread(Thread.currentThread())) {
                 submitSlowPath0(ioOps, seq, userData);
+                return seq;
             } else {
                 executor.execute(() -> submitSlowPath0(ioOps, seq, userData));
+                return seq;
             }
-            return seq;
         }
 
         private void submitFastPath0(IoUringIoOps ioOps, long seq) {
@@ -665,6 +666,168 @@ public final class IoUringIoHandler implements IoHandler {
                 removeLater = false;
                 remove();
             }
+        }
+    }
+
+    private static final class PendingOpSlots {
+        private int[] registrationIds;
+        private byte[] ops;
+        private long[] userDatas;
+        private long[] activeSequences;
+        private int mask;
+        private final AtomicLong nextSequence = new AtomicLong(3);
+
+        PendingOpSlots(int initialCapacity) {
+            int capacity = normalizeCapacity(initialCapacity);
+            registrationIds = new int[capacity];
+            ops = new byte[capacity];
+            userDatas = new long[capacity];
+            activeSequences = new long[capacity];
+            mask = capacity - 1;
+        }
+
+        long nextToken() {
+            long sequence = nextSequence.getAndIncrement();
+            if (sequence <= 0) {
+                throw new IllegalStateException("slow path sequence overflow");
+            }
+            return token(sequence);
+        }
+
+        void registerNormal(long token, int registrationId, byte op, long userData) {
+            long sequence = tokenSequence(token);
+            int slot = ensureWritableSlot(sequence);
+            registrationIds[slot] = registrationId;
+            ops[slot] = op;
+            userDatas[slot] = userData;
+            activeSequences[slot] = sequence;
+        }
+
+        int findSlot(long token) {
+            long sequence = tokenSequence(token);
+            if (sequence == INVALID_ID) {
+                return -1;
+            }
+            int slot = slot(sequence, mask);
+            return activeSequences[slot] == sequence ? slot : -1;
+        }
+
+        int registrationId(int slot) {
+            return registrationIds[slot];
+        }
+
+        byte op(int slot) {
+            return ops[slot];
+        }
+
+        long userData(int slot) {
+            return userDatas[slot];
+        }
+
+        void release(long token) {
+            int slot = findSlot(token);
+            if (slot != -1) {
+                release(slot);
+            }
+        }
+
+        void release(int slot) {
+            registrationIds[slot] = 0;
+            ops[slot] = 0;
+            userDatas[slot] = 0;
+            activeSequences[slot] = INVALID_ID;
+        }
+
+        private int ensureWritableSlot(long sequence) {
+            int mask = this.mask;
+            int slot = slot(sequence, mask);
+            while (activeSequences[slot] != INVALID_ID) {
+                resize();
+                mask = this.mask;
+                slot = slot(sequence, mask);
+            }
+            return slot;
+        }
+
+        private void resize() {
+            int oldCapacity = activeSequences.length;
+            int newCapacity = oldCapacity << 1;
+            if (newCapacity <= 0) {
+                throw new IllegalStateException("slow path table overflow");
+            }
+
+            int[] oldRegistrationIds = registrationIds;
+            byte[] oldOps = ops;
+            long[] oldUserDatas = userDatas;
+            long[] oldActiveSequences = activeSequences;
+
+            int[] newRegistrationIds = new int[newCapacity];
+            byte[] newOps = new byte[newCapacity];
+            long[] newUserDatas = new long[newCapacity];
+            long[] newActiveSequences = new long[newCapacity];
+
+            for (int i = 0; i < oldCapacity; i++) {
+                long sequence = oldActiveSequences[i];
+                if (sequence == INVALID_ID) {
+                    continue;
+                }
+                int newSlot = (sequence & oldCapacity) == 0 ? i : i + oldCapacity;
+                newRegistrationIds[newSlot] = oldRegistrationIds[i];
+                newOps[newSlot] = oldOps[i];
+                newUserDatas[newSlot] = oldUserDatas[i];
+                newActiveSequences[newSlot] = sequence;
+            }
+
+            registrationIds = newRegistrationIds;
+            ops = newOps;
+            userDatas = newUserDatas;
+            activeSequences = newActiveSequences;
+            mask = newCapacity - 1;
+        }
+
+        /**
+         * sequence is always > 0, so its top bit (bit63) is always 0.
+         * sequence layout:
+         *   [ 0 | bit62 ... bit32 | bit31 | bit30 ... bit0 ]
+         * upperBits = sequence >>> 31:
+         *   [ 0 | bit62 ... bit32 | bit31 ]
+         * lowerBits = ((int) sequence & Integer.MAX_VALUE) | Integer.MIN_VALUE:
+         *   [ 1 | bit30 ... bit0 ]   // negative int, high bit is used as a token marker
+         * Final token layout:
+         *   [ 0 | bit62 ... bit32 | bit31 ][ 1 | bit30 ... bit0 ]
+         * @param sequence original sequence
+         * @return real userData
+         */
+        private static long token(long sequence) {
+            // We intentionally do not handle sequence wrap-around here.
+            // `nextSequence` would need to reach Long.MAX_VALUE and overflow,
+            // which is considered practically impossible in this context.
+            long upperBits = sequence >>> 31;
+            int lowerBits = ((int) sequence & Integer.MAX_VALUE) | Integer.MIN_VALUE;
+            return (upperBits << Integer.SIZE) | (lowerBits & 0xFFFFFFFFL);
+        }
+
+        private static long tokenSequence(long token) {
+            int lowerBits = (int) token;
+            if (lowerBits >= 0) {
+                return INVALID_ID;
+            }
+            return (token >>> Integer.SIZE) << 31 | (lowerBits & Integer.MAX_VALUE);
+        }
+
+        private static int slot(long sequence, int mask) {
+            return (int) sequence & mask;
+        }
+
+        private static int normalizeCapacity(int requestedCapacity) {
+            int capacity = 1;
+            while (capacity < requestedCapacity) {
+                capacity <<= 1;
+                if (capacity <= 0) {
+                    throw new IllegalArgumentException("requestedCapacity overflow");
+                }
+            }
+            return capacity;
         }
     }
